Add auth and lawyers routes files
