+++ src/infra/routes.ts
@@ -1,22 +1,10 @@
 import express from 'express';
-import { container } from 'tsyringe';
-import { CreateNormalRoleLawyerController } from './http/controllers/create-normal-role-lawyer';
-import { LoginLawyerController } from './http/controllers/login-lawyer';
 import { documentsRoutes } from './routes/documents';
+import { authRoutes } from './routes/auth';
+import { lawyersRoutes } from './routes/lawyers';
 
 export const routes = express.Router();
 
-const createNormalRoleLawyerController = container.resolve(
-  CreateNormalRoleLawyerController
-);
-const loginLawyerController = container.resolve(LoginLawyerController);
-
+routes.use('/auth', authRoutes);
 routes.use('/documents', documentsRoutes);
-
-routes.post('/lawyers', (request, response, next) =>
-  createNormalRoleLawyerController.handle(request, response, next)
-);
-
-routes.post('/login', (request, response, next) =>
-  loginLawyerController.handle(request, response, next)
-);
+routes.use('/lawyers', lawyersRoutes);
